+++ NEWS.md
@@ -1,6 +1,7 @@
 # moderndive 0.2.0.9000
 
-
+* Added ID argument to get_regression_points() to return a column that ID's the 
+observational units
 
 # moderndive 0.2.0